+++ common/src/main/java/dev/juaanp/barehanded/config/ClientConfig.java
@@ -18,7 +18,7 @@ public class ClientConfig {
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
     private static final File FILE = Paths.get("config", Constants.MOD_ID + "-client.json").toFile();
 
-    public int configVersion = 10;
+    public int configVersion = 11;
 
     public double verticalRotationSensitivity = 0.1;
     public double horizontalRotationSensitivity = 0.1;