+++ common/src/main/java/dev/juaanp/barehanded/network/AssemblyStateC2SPacket.java
@@ -8,7 +8,7 @@ import net.minecraft.resources.ResourceLocation;
 
 public record AssemblyStateC2SPacket(boolean active) implements CustomPacketPayload {
     public static final Type<AssemblyStateC2SPacket> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, "assembly_state_c2s"));
-    
+
     public static final StreamCodec<FriendlyByteBuf, AssemblyStateC2SPacket> CODEC = StreamCodec.of(
             (buf, packet) -> buf.writeBoolean(packet.active()),
             buf -> new AssemblyStateC2SPacket(buf.readBoolean())