+++ common/src/main/java/dev/juaanp/barehanded/physics/GrabTetherSystem.java
@@ -1,35 +1,34 @@
 package dev.juaanp.barehanded.physics;
 
 import dev.juaanp.barehanded.config.ServerConfig;
+import dev.juaanp.barehanded.util.EncumbranceHelper;
 import dev.ryanhcode.sable.Sable;
-import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.phys.Vec3;
 import org.joml.Vector3d;
 
 public class GrabTetherSystem {
+
     public static void applyPhysicalTether(Player player, GrabSession grab, double tension, double actualMaxForce) {
         if (!ServerConfig.INSTANCE.enablePhysicalTether || player.isCreative() || player.isSpectator()) return;
 
         double stretch = tension - grab.distance;
         double armStretchTolerance = ServerConfig.INSTANCE.armStretchTolerance;
-
         double mass = grab.subLevel.getMassTracker().getMass();
-        double objectWeight = mass * ServerConfig.INSTANCE.physicsGravity;
-        double weightRatio = Mth.clamp(objectWeight / actualMaxForce, 0.0, 1.0);
-        double encumbrance = Math.min(Math.pow(weightRatio, 2.0), 1.0);
+
+        double encumbrance = EncumbranceHelper.getEncumbranceRatio(mass, player);
 
         Vector3d currentActualGrabBlockPos = grab.subLevel.logicalPose().transformPosition(new Vector3d(grab.localPivot));
         Vec3 blockPos = new Vec3(currentActualGrabBlockPos.x, currentActualGrabBlockPos.y, currentActualGrabBlockPos.z);
         Vec3 pullDirection = blockPos.subtract(player.getEyePosition()).normalize();
         Vec3 awayDirection = pullDirection.scale(-1.0);
+
         Vec3 currentVel = player.getDeltaMovement();
         Vec3 newVel = currentVel;
 
         if (stretch > armStretchTolerance) {
             double activeStretch = stretch - armStretchTolerance;
             double tetherStiffness = ServerConfig.INSTANCE.tetherStiffnessBase + (ServerConfig.INSTANCE.tetherStiffnessMultiplier * encumbrance);
-
             Vec3 correction = pullDirection.scale(activeStretch * tetherStiffness);
             double correctionY = correction.y < 0 ? correction.y : (correction.y * ServerConfig.INSTANCE.tetherVerticalSmoothing);